Now pulls MiniCodes from firestore
