+++ app/src/main/java/com/example/lab_4_codecatchers/FireStoreActivity.java
@@ -1,21 +1,15 @@
 package com.example.lab_4_codecatchers;
 
-import static android.content.ContentValues.TAG;
-
 import android.util.Log;
 
-import androidx.annotation.NonNull;
-
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QuerySnapshot;
 
 import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Storing and fetching data from firestore database
@@ -66,7 +60,7 @@ public class FireStoreActivity {
     public Task<Void> updateCodes(MiniCode code) {
         return codeCollection
                 .document(code.getHash())
-                .update("locPic", code.getImage(), "location", code.getLocation(), "scannedBy", code.getPlayersWhoScanned());
+                .update("locPic", code.getLocPic(), "location", code.getLocation(), "playersWhoScanned", code.getPlayersWhoScanned());
     }
 
     public Task<Void> removeCode(MiniCode code) {
@@ -92,5 +86,28 @@ public class FireStoreActivity {
 
     public Task<QuerySnapshot> getCodes() {
         return codeCollection.get();
+    }
+
+    protected void fillQRList(){
+        getCodes()
+                .addOnSuccessListener(queryDocumentSnapshots -> {
+                    List<DocumentSnapshot> QRCodes = queryDocumentSnapshots.getDocuments();
+                    if(QRCodes.isEmpty()){
+                        list.setCodes(new ArrayList<MiniCode>());
+                    }else{
+                        ArrayList<MiniCode> codeList = new ArrayList<>();
+                        for(int i = 0; i<QRCodes.size(); i++) {
+                            MiniCode code = QRCodes.get(i).toObject(MiniCode.class);
+                            assert code != null;
+                            MiniCode toAdd = new MiniCode();
+                            toAdd.setHash(code.getHash());
+                            toAdd.setLocPic(code.getLocPic());
+                            toAdd.setLocation(code.getLocation());
+                            toAdd.setPlayersWhoScanned(code.getPlayersWhoScanned());
+                            codeList.add(toAdd);
+                        }
+                        list.setCodes(codeList);
+                    }
+                });
     }
 }
